backend/requirements.txt

fastapi
torch
torchvision
pillow
uvicorn
